+++ source/PlaceFinder/PlaceFinderTest/TestSearchRequestResources.cs
@@ -18,7 +18,7 @@ namespace GeodataStyrelsen.ArcMap.PlaceFinderTest
             string resourceString = searchRequestResources.GetResourceString;
 
             //Assert
-            Assert.AreEqual("Adresser,Stednavne", resourceString);
+            Assert.AreEqual("adresse,stednavn", resourceString);
         }
 
         [Test]
@@ -27,26 +27,24 @@ namespace GeodataStyrelsen.ArcMap.PlaceFinderTest
             //Arrange
             var searchRequestResources = new SearchRequestResources();
             searchRequestResources.Addresses = true;
-            searchRequestResources.CadastralNumber = true;
+            searchRequestResources.Cadastre = true;
+            searchRequestResources.CadastreDeprecated = true;
             searchRequestResources.ElectoralDistrict = true;
             searchRequestResources.HouseNumber = true;
             searchRequestResources.JurisdictionsDistrict = true;
             searchRequestResources.Municipalities = true;
             searchRequestResources.Parish = true;
             searchRequestResources.PlaceNames = true;
-            searchRequestResources.PlaceNames_v2 = true;
-            searchRequestResources.PlaceNames_v3 = true;
             searchRequestResources.PoliceDistrict = true;
             searchRequestResources.PostDistricts = true;
             searchRequestResources.Regions = true;
-            searchRequestResources.Sogne = true;
             searchRequestResources.Street = true;
 
             //Act
             string resourceString = searchRequestResources.GetResourceString;
 
             //Assert
-            Assert.AreEqual("Adresser,Veje,Husnumre,Kommuner,Matrikelnumre,Stednavne,Stednavne_v2,Stednavne_v3,Opstillingskredse,Politikredse,Postdistrikter,Regioner,Retskredse,Sogne", resourceString);
+            Assert.AreEqual("adresse,husnummer,kommune,matrikel,matrikel_udgaaet,navngivenvej,opstillingskreds,politikreds,postnummer,region,retskreds,sogn,stednavn", resourceString);
         }
     }
 }