+++ changelog.md
@@ -3,6 +3,17 @@ All notable changes to this project will be documented in this file.
 
 ## [v?.?.?] (Unreleased)
 
+## [v2.0.0.1] (2024-02-14)
+
+### Changed
+- Hits resource sorting and concurrentbag data structure reordered results issue fixed
+
+## [v2.0.0] (2024-01-30)
+
+### Changed
+- Using the new gsearch service from dataforsyningen.dk at https://api.dataforsyningen.dk/rest/gsearch/v2.0
+- Response parsing using regex operations to avoid challenge with loading libraries under ArcMap 10.8.1
+
 ## [v1.5.0] (2016-10-26)
 
 ### Changed